### 2016-12-05 #122 @arfon
@biorelated - would you be willing to edit this submission for JOSS?
