### 2016-12-06 9515f84
Sync version in codemeta with github, fix some paper typos
