+++ paper/paper.md
@@ -28,10 +28,9 @@ finding an expect-value cutoff per alignment length, and then selecting non-reci
 alignments which meet the minimum threshold.
 
 Unfortunately, the original implementation uses the relatively slow NCBI BLAST+ 
-[altschul_basic_1990], and is implemented in Ruby, which requires users to leave 
-the Python-dominated bioinformatics software
-system. shmlast makes CRBH available to users in Python, while also greatly improving
-performance by using the LAST aligner [@kielbasa_adaptive_2011} for initial homology searches.
+[@altschul_basic_1990], and is implemented in Ruby, which requires users to leave 
+the Python-dominated bioinformatics ecosystem. shmlast makes CRBH available to users in Python, while also greatly improving
+performance by using the LAST aligner [@kielbasa_adaptive_2011] for initial homology searches.
 Other improvements include outputting the list of cutoffs generated by its model along with a plot
 of the decision boundary to aid in quality control, as well as using gnu-parallel [@tange_gnu_2011]
 to parallelize execution across multiple cores or nodes in a cluster environment.
