### 2016-12-17 #122 @biorelated
Sorry @genomematt  for the delay i assigning the review task. I was away from Github the last 2 days. Glad you can review this!
