### 2016-12-18 #142 @whedon
[REVIEW]: shmlast: An improved implementation of Conditional Reciprocal Best Hits with LAST and Python
**Submitting author:** @camillescott (<a href="http://orcid.org/0000-0001-8822-8779">Camille Scott</a>)
**Repository:** <a href="https://github.com/camillescott/shmlast" target ="_blank">https://github.com/camillescott/shmlast</a>
**Version:** v1.1
