### 2017-01-04 #142 @genomematt
I am still working my way through this, but some initial comments @camillescott.

I think that you need a more from first principles statement of need that outlines the sequence matching problem that CRBH is addressing (remembering that this is a generalist scientific computing journal) so that readers don't need to chase a reference to get a basic idea of what the software does.

You need contribution, bug report and support guidelines - check out some other JOSS articles for what this should look like.

A real world (ish) worked example, ideally as a jupyter notebook or doctest file would greatly enhance, and would help with explaining the statement of need.

Very pleased to see coverage as part of your testing.  Your exclusions all look very sensible on first pass.  You may wish to add `#pragma no cover` directives marking the code you don't test.  This makes it easier to track in your code and makes full coverage 100%. (suggestion not a requirement)

As you make a claim of improved performance (speed) it would be good to quantify that improvement.

I had an install issue but I am pretty sure it was related to a prior install not yours.  I will redo on a clean VM to confirm.
