### 2017-01-09 #142 @camillescott
Thanks @genomematt. I'll add a section to the paper describing the approach in a bit more detail. I can easily add a figure with performance comps as well.
