### 2017-01-09 7aef99d
Add methods section
