+++ paper/paper.md
@@ -35,6 +35,34 @@ Other improvements include outputting the list of cutoffs generated by its model
 of the decision boundary to aid in quality control, as well as using gnu-parallel [@tange_gnu_2011]
 to parallelize execution across multiple cores or nodes in a cluster environment.
 
-![A CRBH model generated by shmlast between *Schizosaccharomyces pombe* and *Rattus norvegicus*. Hits above the blue dotted line will be kept.](sacPom.cdna.fa.x.Rattus_norvegicus.UP000002494.ref.pep.fa.crbl.model.plot.pdf)
+# Methods
+
+RBH is a relatively old method for determining orthologs between two sequence databases. Orthology
+is distinguished from sequence similarity by descent; while two sequences with high similarity are
+likely to have structural of functional homology, they are orthologous if they
+share a common ancestor sequence. It is difficult to know for sure whether two sequences are
+orthologs, but orthologous groups of sequences are critical for a variety of analyses surrounding
+structure, function, and evolution, and particularly, for annotation. As such, many methods have 
+been developed for separating high-similarity alignments from their orthologous counterparts. RBH is the regal elder of these
+methods, and although it is simplistic compared to newer clustering and graph-based methods, it
+remains in wide use due to its low false-positive rate and ease of implementation. It is performed
+as follows: given two sets of sequences $A$ and $B$, sequences $a_i \in A$ and $b_j \in B$ are
+Reciprocal Best Hits if $b_j$ has the highest scoring sequence alignment in $B$ for $a_i$ and $a_i$
+has the highest scoring sequence alignment in $A$ for $b_j$. $a_i$ and $b_j$ then have a high
+probability of being orthologs.
+
+While this method works well for finding orthologs between two sets of proteins from different
+species, it is less effective for annotating newly assembled transcriptomes from existing protein databases.
+Transcriptomes are confounded by alternative splicing, causing several transcripts to share
+subsequences, which may prevent RBH detection between a translated transcript and its protein, even
+when an orthology relationship exists. Aubry et al., and this implementation, circumvent that
+problem by first using the reciprocals to establish a score cutoff for each alignment length, and
+then keeping *any* alignment which passes that cutoff. This prevents alignments with
+high-likelihood of being orthologs based on sequence identity from being discarded due to the high
+specificity of RBH.
+
+
+
+![A CRBH model generated by shmlast between *Schizosaccharomyces pombe* and *Rattus norvegicus*. Hits with scores above the blue dotted line will be kept.](sacPom.cdna.fa.x.Rattus_norvegicus.UP000002494.ref.pep.fa.crbl.model.plot.pdf)
 
 # References
