### 2017-01-09 6085838
Add contribution guidelines
