+++ CONTRIBUTING.md
@@ -1,0 +1,7 @@
+Contributions are welcome! Please file issues in the tracker for bugs, feature requests, or general
+questions or difficulties.
+
+I use
+[GitHub Flow style development](https://guides.github.com/introduction/flow/).
+Please fork the repository and set up a pull request against the master branch with any changes
+you want me to consider merging.
