### 2017-01-10 e34f51b
Add perf info
