### 2017-01-10 587f593
Add perf database citations and figure
