+++ paper/paper.bib
@@ -9,13 +9,31 @@
 	language = {en},
 	number = {3},
 	urldate = {2015-10-02},
-	journal = {Genome Res.},
+	journal = {Genome Research},
 	author = {Kiełbasa, Szymon M. and Wan, Raymond and Sato, Kengo and Horton, Paul and Frith, Martin C.},
 	month = mar,
 	year = {2011},
 	pmid = {21209072},
 	pages = {487--493},
-	file = {Full Text PDF:/Users/camille/Library/Application Support/Zotero/Profiles/oy306j3r.default/zotero/storage/ZIU7D72H/Kiełbasa et al. - 2011 - Adaptive seeds tame genomic sequence comparison.pdf:application/pdf;Snapshot:/Users/camille/Library/Application Support/Zotero/Profiles/oy306j3r.default/zotero/storage/RCGHWSXW/487.html:text/html}
+	file = {Full Text PDF:/home/camille/.zotero/zotero/kwx0so0v.default/zotero/storage/ZIU7D72H/Kiełbasa et al. - 2011 - Adaptive seeds tame genomic sequence comparison.pdf:application/pdf;Snapshot:/home/camille/.zotero/zotero/kwx0so0v.default/zotero/storage/RCGHWSXW/487.html:text/html}
+}
+
+@article{aubry_deep_2014,
+	title = {Deep {Evolutionary} {Comparison} of {Gene} {Expression} {Identifies} {Parallel} {Recruitment} of {Trans} -{Factors} in {Two} {Independent} {Origins} of {C} 4 {Photosynthesis}},
+	volume = {10},
+	issn = {1553-7404},
+	url = {http://journals.plos.org/plosgenetics/article?id=10.1371/journal.pgen.1004365},
+	doi = {10.1371/journal.pgen.1004365},
+	abstract = {Author Summary   C 4  photosynthesis is one of the most successful and widespread examples of convergent evolution; the first C 4  plant evolved long after the extinction of the dinosaurs, yet C 4  species now account for ∼30\% of primary productivity on earth. Compared with ancestral C 3  photosynthesis, the C 4  pathway allows faster rates of growth, and thus international efforts have been mustered to introduce advantageous C 4  traits into important C 3  crops to increase their yield. However, the transition from C 3  to C 4  involves complex alterations to leaf anatomy and biochemistry. Despite these multiple changes, C 4  photosynthesis has evolved independently at least 60 times. Through DNA and RNA sequencing we are beginning define a catalog of genes associated with C 3  or C 4  photosynthesis. However, we know little about how these genes act co-ordinately to bring about the convergent C 4  phenotype. In this work we develop a new informatics framework to reveal that two independent lineages of C 4  plants have co-opted the same regulators of gene expression to generate the C 4  leaf. Our findings provide a new paradigm for investigating the genetics of convergent traits and the origin of convergent phenotypes. Moreover, they reveal significant new insight into the regulatory mechanisms governing the origins of C 4  photosynthesis.},
+	number = {6},
+	urldate = {2016-05-11},
+	journal = {PLOS Genet},
+	author = {Aubry, Sylvain and Kelly, Steven and Kümpers, Britta M. C. and Smith-Unna, Richard D. and Hibberd, Julian M.},
+	month = jun,
+	year = {2014},
+	keywords = {C4 photosynthesis, Gene expression, Gene regulation, Leaves, Maize, Rice, Transcription factors, Transcriptome analysis},
+	pages = {e1004365},
[…]
