+++ paper/paper.md
@@ -61,6 +61,14 @@ then keeping *any* alignment which passes that cutoff. This prevents alignments 
 high-likelihood of being orthologs based on sequence identity from being discarded due to the high
 specificity of RBH.
 
+# Performance
+
+shmlast benefits immensely from the use of LAST over BLAST. It scales well by using gnu-parallel,
+and can be distributed across clusters for particularly large runs.
+
+![Performance comparison with [*Schizosaccharomyces pombe*](ftp://ftp.ebi.ac.uk/pub/databases/pombase/FASTA/cdna_nointrons_utrs.fa.gz) as the query transcriptome and
+[Nematostella vectensis](ftp://ftp.ebi.ac.uk/pub/databases/reference_proteomes/QfO/Eukaryota/UP000001593_45351.fasta.gz) as the target proteome.](../perf/results.pdf)
+
 
 
 ![A CRBH model generated by shmlast between *Schizosaccharomyces pombe* and *Rattus norvegicus*. Hits with scores above the blue dotted line will be kept.](sacPom.cdna.fa.x.Rattus_norvegicus.UP000002494.ref.pep.fa.crbl.model.plot.pdf)
