### 2017-01-11 72bb954
Add JOSS test command
