+++ paper/paper.md
@@ -71,6 +71,6 @@ and can be distributed across clusters for particularly large runs.
 
 
 
-![A CRBH model generated by shmlast between *Schizosaccharomyces pombe* and *Rattus norvegicus*. Hits with scores above the blue dotted line will be kept.](sacPom.cdna.fa.x.Rattus_norvegicus.UP000002494.ref.pep.fa.crbl.model.plot.pdf)
+![CRBH model generated from the performance comparison. Hits with scores above the blue dotted line will be kept.](sacPom.cdna.fa.x.Nematostella_vectensis.uniprot.pep.fa.crbl.model.plot.pdf)
 
 # References
