### 2017-01-11 #142 @camillescott
Update: I have added a performance comparison figure, a section describing the method in a bit more detail, and contribution guidelines.
