### 2017-01-14 #142 @genomematt
Ok @camillescott the added bits are good.  My install issue was a conda problem and all was fine on a clean vm.

Given that there are multiple dependencies I think there should be a test step at the end of your install procedure.  This would not need to be comprehensive testing of code, just a 'if you have installed it correctly this input will give this output.'

I am giving you a scraping pass on the real world application documentation, but an additional notebook  that does not hide the actual running in a script like the performance notebook, showing real transcripts against a real db and then interpreting the csv to answer the biological question would enhance.
As it stands the documentation is good for anyone who has done this sort of procedure before, and not as easy for someone who has a set of transcripts and is asking 'what is the best way to match these to the most closely related model organism'

Neither of those would stop me recommending accepting, but the complete lack of community guidelines does.
```Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support```
