### 2017-01-17 #142 @camillescott
What is expected for guidelines? I did add a CONTRIBUTING.md: https://github.com/camillescott/shmlast/blob/master/CONTRIBUTING.md

I followed the example of several other accepted projects on that -- should I add more?
