### 2017-01-17 85edd24
Add link to contribution guidelines to README
