### 2017-01-17 7a100b4
Update README.rst from README.md for PyPI
