### 2017-01-17 #142 @genomematt
Sorry @camillescott I missed the CONTRIBUTING.md.  I suspect a lot of other people especially those with bugs or support questions would as well.  At least link it from the readme, and personally I would be inclined to put such a short section just in the readme.
