### 2017-01-17 #142 @genomematt
Ok @biorelated I consider this ready to accept.
