### 2017-01-18 #142 @camillescott
Copy that @genomematt -- added it to the readme :)
