### 2017-01-18 9ff9a4c
Update contribution guidelines with information to submit with issues
