+++ CONTRIBUTING.md
@@ -1,5 +1,11 @@
 Contributions are welcome! Please file issues in the tracker for bugs, feature requests, or general
-questions or difficulties.
+questions or difficulties. When filing issues, please include, along with a description of the
+problem:
+
+* Operating system (distribution and version)
+* Your python environment (either `pip freeze` or `conda list`)
+* Your LAST version `lastal --version`
+* If applicable, the output produced by the failed run.
 
 I use
 [GitHub Flow style development](https://guides.github.com/introduction/flow/).
