### 2017-01-18 43ab51c
Add issue template
