+++ ISSUE_TEMPLATE.md
@@ -1,0 +1,13 @@
+Please see the [contribution guidelines](CONTRIBUTING.md) for details on submitting an issue.
+
+## Problem Description
+
+## My Environment
+
+### Operating System
+
+### Python Environment
+
+### LAST Version
+
+### Shmlast Output
