### 2017-01-18 7ce3d95
tweak issue template format
