### 2017-01-18 fd4f059
Github already links to CONTRIBUTING.md, remove link from template
