+++ ISSUE_TEMPLATE.md
@@ -1,5 +1,3 @@
-Please see the CONTRIBUTING.md for details on submitting an issue.
-
 ## Problem Description
 
 ## My Environment
