### 2017-01-27 #142 @camillescott
@arfon: I've archived on Zenodo; here is the DOI: http://doi.org/10.5281/zenodo.260437

note that the final published version after the requested changes is v1.1, not v1.0.2 as in Whedon's first post in this issue
